+++ src/Molecules/Input/Text/Text.types.ts
@@ -5,6 +5,7 @@ export type Props = {
   /** But you can hide it visually */
   hideLabel?: boolean;
   labelTooltip?: string;
+  autoComplete?: string;
   autoFocus?: boolean;
   name?: string;
   error?: string;