+++ src/Molecules/Modal/Modal.types.ts
@@ -5,10 +5,17 @@ export type InnerProps = {
   title?: React.ReactNode;
   footer?: React.ReactNode;
   hideClose?: boolean;
+  /** @default false */
+  closeOnBackdropClick?: boolean;
+};
+
+export type BackdropProps = {
+  onClick: (e: React.ChangeEvent<HTMLElement>) => void;
 };
 
 export type DialogProps = {
   show: boolean;
+  onClick: (e: React.MouseEvent<HTMLElement>) => void;
 };
 
 export type Props = {